+++ contracts/attrs/src/helpers.rs
@@ -1,4 +1,6 @@
-use cosmwasm_std::{from_binary, to_binary, Addr, Binary, CosmosMsg, Empty, StdError, StdResult};
+use cosmwasm_std::{
+    from_json, to_json_binary, Addr, Binary, CosmosMsg, Empty, StdError, StdResult,
+};
 use provwasm_std::types::provenance::attribute::v1::{
     AttributeQuerier, AttributeType, MsgAddAttributeRequest, MsgDeleteAttributeRequest,
     MsgDeleteDistinctAttributeRequest, MsgUpdateAttributeRequest,
@@ -67,7 +69,7 @@ pub fn add_json_attribute<H: Into<Addr>, S: Into<String>, T: Serialize + ?Sized>
         address,
         contract_address,
         name,
-        to_binary(data)?,
+        to_json_binary(data)?,
         AttributeType::Json,
     )
 }
@@ -149,7 +151,7 @@ pub fn get_json_attributes<H: Into<Addr>, S: Into<String>, T: DeserializeOwned>(
     resp.attributes
         .iter()
         .filter(|a| a.attribute_type == AttributeType::Json as i32)
-        .map(|a| from_binary(&Binary::from(a.value.clone())))
+        .map(|a| from_json(&Binary::from(a.value.clone())))
         .collect()
 }