+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Create security contract to link assets with a security
   type ([#138](https://github.com/provenance-io/provwasm/issues/138))
 * Bump to Provenance 1.18.0-rc2 ([#143](https://github.com/provenance-io/provwasm/issues/143))
+* remove nft, security, and template contracts ([#145](https://github.com/provenance-io/provwasm/pull/145))
+* Bump to Provenance 1.18.0 and add new exchange types ([#146](https://github.com/provenance-io/provwasm/issues/146))
+* Update dependencies, including cosmwasm 1.5 ([#147](https://github.com/provenance-io/provwasm/issues/147))
 
 ## Releases